feat: add leaveButton to WaitingRoom and refactor into own component
